Moved changeIcon() to HudItemCrosshairs
changeIcon() is used to change the sprite for HudItemIcon. However, this
would not be often used outside of the class in any situation other the
Crosshairs.
