+++ src/main/java/net/binaryaura/customize/common/Customize.java
@@ -9,7 +9,6 @@ import net.minecraftforge.client.GuiIngameForge;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
 import net.minecraftforge.fml.common.Mod.Instance;
-import net.minecraftforge.fml.common.ModContainer.Disableable;
 import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLModDisabledEvent;
@@ -20,7 +19,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class Customize {
     public static final String MODID = "customize";
 	public static final String NAME = "Customize";
-    public static final String VERSION = "1A1-5";
+    public static final String VERSION = "1A1-6";
     public static final String CLIENTPROXY = "net.binaryaura.customize.client.";
     public static final String COMMONPROXY = "net.binaryaura.customize.common.";
